chunkky update

+++ libft/ft_bzero.c
@@ -1,5 +1,6 @@
-void	*ft_memset(void *b, int c, int len);
-void	ft_bzero(void *s, int n)
+#include "libft.h"
+void	*ft_memset(void *b, int c, size_t len);
+void	ft_bzero(void *s, size_t n)
 {
 	ft_memset(s, 0, n);
 }
